ml/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
pydantic==2.9.2
python-multipart==0.0.12
xgboost
joblib
torch
torchvision
Pillow 